Update reszoom to match my personal requirements
reszoom originally is meant to set the zoom to twice the size on HiDPI
displays while it set it to normal (100%) size on others.

However in my personal setup this does not apply. I do not want my
external 28-inch 4K display just to be a beautiful retina 2K display.

But I want to use reszoom to switch zoom based on which screen I
currently use. On the internal screen of my MacBook I normally want 100%
zoom (except I set the display to native resolution of 2560x1600) while
on the external screen I want the zoom to be increased a bit to 125%.